Mudando somente o ip da url (o restante fica igual)

+++ app/get-custom-rtsp-camera.sh
@@ -1,5 +1,7 @@
 params=$1
 
+USAR_NGROK=false
+
 function get_config {
   config=$(jq -r '' "config.json")
   echo $config | jq '.'$1''
@@ -21,14 +23,32 @@ CAMERA_ADRESS_IP=$(get_params "camera_adress_ip")
 CAMERA_ADRESS_PORTA=$(get_params "camera_adress_porta")
 CAMERA_CHANNEL=$(get_params "camera_channel")
 
-function get_ngrok_obj {
-  resp=$(get_config "ngrok_tunnel")
+function get_sc_tunnel_obj {
+  resp=$(get_config "sc_tunnel")
 
   if [[ $resp == null ]]; then
-    resp=$(ngrok api tunnels list | jq '.tunnels[0]')
+    # Com NGROK
+    if [[ $USAR_NGROK == true ]]; then
+      echo "com ngrok"
+      resp=$(ngrok api tunnels list | jq '.tunnels[0]')
+    else
+      # Com sctunnel
+      pid=$(ps aux | grep "$porta:$CAMERA_ADRESS_IP:$CAMERA_ADRESS_PORTA ubuntu@23.22.12.192" | awk '{print $2}')
+      $(kill -9 $pid > /dev/null &)
+
+      porta=$(ssh -i "~/portaria_staging_ssh_pem_key.pem" ubuntu@23.22.12.192 'bash -s' < find_unused_port.sh)
+      ssh -N -o ServerAliveInterval=20 -i "~/portaria_staging_ssh_pem_key.pem" -R $porta:$CAMERA_ADRESS_IP:$CAMERA_ADRESS_PORTA ubuntu@23.22.12.192 > /dev/null &
+
+      public_url="23.22.12.192:$porta"
+      forwards_to="$CAMERA_ADRESS_IP:$CAMERA_ADRESS_PORTA"
+      resp="{\"public_url\": \"$public_url\", \"forwards_to\": \"$forwards_to\"}"
+    fi
 
     # salvando nova configuração
-    new_config=$(get_config | jq --argjson v "$resp" '.ngrok_tunnel = $v')
+    config=$(get_config | jq '')
+    new_config=$(echo "{}" | jq --argjson v "$resp" '.sc_tunnel = $v')
+
+    new_config=$(echo ''$config' '$new_config'' | jq -s add)
     echo $new_config | jq '.' > config.json
   fi
 
@@ -37,7 +57,13 @@ function get_ngrok_obj {
 
 function mount_current_url {
   current_url=$(echo $1 | jq '.public_url' | tr -d '"' | sed -e "s/tcp:[/][/]//")
-  current_url="rtsp://$CAMERA_USER:$CAMERA_USER_PASS@$current_url/cam/realmonitor?channel=$CAMERA_CHANNEL&subtype=1"
+
+  url_controller=$(get_params "camera_adress_url_controller")
+  if [[ url_controller == null ]]; then
+    url_controller="cam/realmonitor?channel=$CAMERA_CHANNEL&subtype=1"
+  fi
+
+  current_url="rtsp://$CAMERA_USER:$CAMERA_USER_PASS@$current_url/$url_controller"
 
   echo $current_url
 }
@@ -48,25 +74,27 @@ function get_current_stream_obj {
   echo $current_str_obj
 }
 
-current_ngrok_obj=$(get_ngrok_obj)
+current_tunnel_obj=$(get_sc_tunnel_obj)
 current_stream_obj=$(get_current_stream_obj)
 
-current_url=$(mount_current_url "$current_ngrok_obj")
+current_url=$(mount_current_url "$current_tunnel_obj")
 current_stream_url=$(echo $current_stream_obj | jq '.payload.url?' | tr -d '"')
 
-if [[ "$current_url" != "$current_stream_url" ]]; then
-  forwards_to=$(echo $current_ngrok_obj | jq '.forwards_to' | tr -d '"')
-  if [[ "$CAMERA_ADRESS_IP:$CAMERA_ADRESS_PORTA" != "$forwards_to" ]]
-  then
-    # Destruindo serviços antigos
-    killall ngrok
+if [[ $USAR_NGROK ]]; then
+  if [[ "$current_url" != "$current_stream_url" ]]; then
+    forwards_to=$(echo $current_tunnel_obj | jq '.forwards_to' | tr -d '"')
+    if [[ "$CAMERA_ADRESS_IP:$CAMERA_ADRESS_PORTA" != "$forwards_to" ]]
+    then
+      # Destruindo serviços antigos
+      killall ngrok
 
-    # Ligando ngrok em background
-    ngrok tcp $CAMERA_ADRESS_IP:$CAMERA_ADRESS_PORTA > /dev/null &
+      # Ligando ngrok em background
+      ngrok tcp $CAMERA_ADRESS_IP:$CAMERA_ADRESS_PORTA > /dev/null &
 
-    current_ngrok_obj=$(get_ngrok_obj)
+      current_tunnel_obj=$(get_sc_tunnel_obj)
 
-    current_url=$(mount_current_url "$current_ngrok_obj")
+      current_url=$(mount_current_url "$current_tunnel_obj")
+    fi
   fi
 fi
